Added a new enhanced way for users to process queries with.
Consists of an AJAX interface to work with BackgrounDRb, so the
user gets information about what's going on all the time, as the
processes could end up taking quite a while. Also only lets one
process run at a time and waits for server load to fall down before
starting new processes, so it doesn't force the server to its knees,
if the users are being impatient.